% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{asymmetric_lp}
\alias{asymmetric_lp}
\title{Apply asymmetric polynomials filters}
\usage{
asymmetric_lp(
  y,
  horizon,
  degree = 3,
  kernel = c("Henderson", "Uniform", "Biweight", "Trapezoidal", "Triweight", "Tricube",
    "Gaussian", "Triangular", "Parabolic"),
  endpoints = c("LC", "QL", "CQ", "CC", "DAF"),
  ic = 4.5,
  q = 0,
  tweight = 0,
  passband = pi/12
)
}
\arguments{
\item{y}{input time-series}

\item{horizon}{horizon of the filter}

\item{degree}{degree of polynomial}

\item{kernel}{kernel uses}

\item{endpoints}{methode for endpoints}

\item{ic}{ic ratio}

\item{q}{number of available data for the assymetric filter.
By default (\code{q = 0}) a real-time filter is used.}

\item{tweight}{timeliness weight}

\item{passband}{passband threshold}
}
\value{
the target signal
}
\description{
Apply asymmetric polynomials filters
}
