% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_filters.R
\name{finite_filters}
\alias{finite_filters}
\alias{show,finite_filters-method}
\alias{*,finite_filters,moving_average-method}
\alias{*,moving_average,finite_filters-method}
\alias{*,finite_filters,numeric-method}
\alias{*,ANY,finite_filters-method}
\alias{*,finite_filters,ANY-method}
\alias{+,finite_filters,numeric-method}
\alias{+,numeric,finite_filters-method}
\alias{+,finite_filters,moving_average-method}
\alias{+,moving_average,finite_filters-method}
\alias{+,finite_filters,missing-method}
\alias{-,finite_filters,missing-method}
\alias{-,finite_filters,moving_average-method}
\alias{-,moving_average,finite_filters-method}
\alias{-,finite_filters,numeric-method}
\alias{-,numeric,finite_filters-method}
\alias{/,finite_filters,numeric-method}
\alias{*,finite_filters,finite_filters-method}
\title{Manipulating Finite Filters}
\usage{
finite_filters(
  sfilter,
  rfilters = NULL,
  lfilters = NULL,
  first_to_last = FALSE
)

\S4method{show}{finite_filters}(object)

\S4method{*}{finite_filters,moving_average}(e1, e2)

\S4method{*}{moving_average,finite_filters}(e1, e2)

\S4method{*}{finite_filters,numeric}(e1, e2)

\S4method{*}{ANY,finite_filters}(e1, e2)

\S4method{*}{finite_filters,ANY}(e1, e2)

\S4method{+}{finite_filters,numeric}(e1, e2)

\S4method{+}{numeric,finite_filters}(e1, e2)

\S4method{+}{finite_filters,moving_average}(e1, e2)

\S4method{+}{moving_average,finite_filters}(e1, e2)

\S4method{+}{finite_filters,missing}(e1, e2)

\S4method{-}{finite_filters,missing}(e1, e2)

\S4method{-}{finite_filters,moving_average}(e1, e2)

\S4method{-}{moving_average,finite_filters}(e1, e2)

\S4method{-}{finite_filters,numeric}(e1, e2)

\S4method{-}{numeric,finite_filters}(e1, e2)

\S4method{/}{finite_filters,numeric}(e1, e2)

\S4method{*}{finite_filters,finite_filters}(e1, e2)
}
\arguments{
\item{sfilter}{the symmetric filter}

\item{rfilters}{the right filters (used on the last points), from the first filter to the last (real-time)}

\item{lfilters}{the left filters (used on the first points).}

\item{first_to_last}{boolean}
}
\description{
Manipulating Finite Filters
}
