% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_filters.R
\name{fstfilter}
\alias{fstfilter}
\title{Estimation of a filter using the Fidelity-Smoothness-Timeliness criteria}
\usage{
fstfilter(
  lags,
  leads,
  pdegree = 2,
  smoothness.weight = 1,
  smoothness.degree = 3,
  timeliness.weight = 0,
  timeliness.passband = pi/6,
  timeliness.antiphase = TRUE
)
}
\arguments{
\item{lags}{Lags of the filter (should be positive).}

\item{leads}{Leads of the filter (should be positive or 0).}

\item{pdegree}{Local polynomials preservation: max degree.}

\item{smoothness.weight}{Weight for the smoothness criterion (in \eqn{[0, 1]}).}

\item{smoothness.degree}{Degree of the smoothness criterion (3 for Henderson).}

\item{timeliness.weight}{Weight for the Timeliness criterion (in \eqn{[0, 1[}). \code{sweight+tweight} should be in \eqn{[0,1]}.}

\item{timeliness.passband}{Passband for the timeliness criterion (in radians). The phase effect is computed in \eqn{[0, passband]}.}

\item{timeliness.antiphase}{boolean indicating if the timeliness should be computed analytically (\code{TRUE}) or numerically (\code{FALSE}).}

\item{passband}{Passband of the filter (only used if tweight is positive).}
}
\value{
The selected filter, its gain and phase, and the values of the 3 criteria.
}
\description{
Estimation of a filter using the Fidelity-Smoothness-Timeliness criteria
}
\examples{
filter <- fstfilter(lags = 6, leads = 0)
filter$filter
filter$criteria
}
\references{
Grun-Rehomme, Michel, Fabien Guggemos, and Dominique Ladiray (2018). “Asymmetric Moving Averages Minimizing Phase Shift”. In: Handbook on Seasonal Adjustment.
}
