% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{get_kernel}
\alias{get_kernel}
\title{Get the coefficients of a kernel}
\usage{
get_kernel(
  kernel = c("Henderson", "Uniform", "Triangular", "Epanechnikov", "Parabolic",
    "Biweight", "Triweight", "Tricube", "Trapezoidal", "Gaussian"),
  horizon
)
}
\arguments{
\item{kernel}{kernel uses}

\item{horizon}{horizon of the filter}
}
\value{
\code{tskernel} object (see \link[stats]{kernel}).
}
\description{
Get the coefficients of a kernel
}
\examples{
get_kernel("Henderson", horizon = 3)
}
