% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_finite_filters.R
\name{impute_last_obs}
\alias{impute_last_obs}
\title{Impute Incomplete Finite Filters}
\usage{
impute_last_obs(x, n, nperiod = 1)
}
\arguments{
\item{x}{a \code{\link[=finite_filters]{finite_filters()}} object.}

\item{n}{integer specifying the number of imputed periods.}

\item{nperiod}{integer specifying how to imput missing date.
\code{nperiod = 1} means imputation using last filtered data (1 period backward),
\code{nperiod = 12} with monthly datameans imputation using last year filted data, etc.}
}
\description{
Impute Incomplete Finite Filters
}
\examples{
y <- window(retailsa$AllOtherGenMerchandiseStores, start = 2008)
M3X3 <- macurves("S3X3")
M2X12 <- (simple_ma(12, -6) + simple_ma(12, -5)) / 2
composite_ma <- M3X3 * M2X12
# The last 6 points cannot be computed
composite_ma
composite_ma * y
# they can be computed using the last filtered data
# e.g. to impute the first 3 missing months with last period:
impute_last_obs(composite_ma, n = 3, nperiod = 1) * y
# or using the filtered data of the same month in previous year
impute_last_obs(composite_ma, n = 6, nperiod = 12) * y
}
