% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{localpolynomials}
\alias{localpolynomials}
\title{Title}
\usage{
localpolynomials(
  y,
  horizon,
  degree = 3,
  kernel = c("Henderson", "Uniform", "Biweight", "Triweight", "Tricube", "Gaussian",
    "Triangular", "Parabolic"),
  endpoints = c("LC", "QL", "CQ", "CC", "DAF"),
  ic = 4.5
)
}
\arguments{
\item{ic}{}
}
\value{

}
\description{
Title
}
