% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{lpp_properties}
\alias{lpp_properties}
\title{Get properties of local polynomials filters}
\usage{
lpp_properties(
  horizon,
  degree = 3,
  kernel = c("Henderson", "Uniform", "Biweight", "Trapezoidal", "Triweight", "Tricube",
    "Gaussian", "Triangular", "Parabolic"),
  endpoints = c("LC", "QL", "CQ", "CC", "DAF", "CN"),
  ic = 4.5,
  tweight = 0,
  passband = pi/12
)
}
\arguments{
\item{horizon}{horizon (bandwidth) of the symmetric filter .}

\item{degree}{degree of polynomial.}

\item{kernel}{kernel uses.}

\item{endpoints}{methode for endpoints.}

\item{ic}{ic ratio.}

\item{tweight}{timeliness weight.}

\item{passband}{passband threshold.}
}
\value{
\code{list} with coefficients, gain and phase values
}
\description{
Get properties of local polynomials filters
}
\details{
\itemize{
\item "LC": Linear-Constant filter
\item "QL": Quadratic-Linear filter
\item "CQ": Cubic-Quadratic filter
\item "CC": Constant-Constant filter
\item "DAF": Direct Asymmetric filter
\item "CN": Cut and Normalized Filter
}
}
\examples{
henderson_f <- lpp_properties(horizon = 6, kernel = "Henderson")
plot_coef(henderson_f)
}
\references{
Proietti, Tommaso and Alessandra Luati (2008). “Real time estimation in local polynomial regression, with application to trend-cycle analysis”.
}
\seealso{
\code{\link{localpolynomials}}
}
