% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average.R
\name{moving_average}
\alias{moving_average}
\alias{is.moving_average}
\alias{is_symmetric}
\alias{get_upper_bound}
\alias{get_lower_bound}
\alias{mirror}
\alias{length.moving_average}
\alias{to_seasonal}
\alias{show,moving_average-method}
\alias{sum.moving_average}
\alias{+,moving_average,moving_average-method}
\alias{+,moving_average,numeric-method}
\alias{+,numeric,moving_average-method}
\alias{+,moving_average,missing-method}
\alias{-,moving_average,missing-method}
\alias{-,moving_average,moving_average-method}
\alias{-,moving_average,numeric-method}
\alias{-,numeric,moving_average-method}
\alias{*,moving_average,moving_average-method}
\alias{*,moving_average,numeric-method}
\alias{*,ANY,moving_average-method}
\alias{*,moving_average,ANY-method}
\alias{*,numeric,moving_average-method}
\alias{/,moving_average,numeric-method}
\title{Manipulation of moving averages}
\usage{
moving_average(x, lags)

is.moving_average(x)

is_symmetric(x)

get_upper_bound(x)

get_lower_bound(x)

mirror(x)

\method{length}{moving_average}(x)

to_seasonal(x, s)

\S4method{show}{moving_average}(object)

\method{sum}{moving_average}(..., na.rm = FALSE)

\S4method{+}{moving_average,moving_average}(e1, e2)

\S4method{+}{moving_average,numeric}(e1, e2)

\S4method{+}{numeric,moving_average}(e1, e2)

\S4method{+}{moving_average,missing}(e1, e2)

\S4method{-}{moving_average,missing}(e1, e2)

\S4method{-}{moving_average,moving_average}(e1, e2)

\S4method{-}{moving_average,numeric}(e1, e2)

\S4method{-}{numeric,moving_average}(e1, e2)

\S4method{*}{moving_average,moving_average}(e1, e2)

\S4method{*}{moving_average,numeric}(e1, e2)

\S4method{*}{ANY,moving_average}(e1, e2)

\S4method{*}{moving_average,ANY}(e1, e2)

\S4method{*}{numeric,moving_average}(e1, e2)

\S4method{/}{moving_average,numeric}(e1, e2)
}
\arguments{
\item{x}{vector of coefficients}

\item{lags}{integer indicating the number of lags of the moving average}

\item{s}{seasonal period for the \code{to_seasonal()} function.}

\item{object, e1, e2}{moving averages}

\item{..., na.rm}{see \link[base]{mean}}
}
\description{
Manipulation of moving averages
}
