% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mse}
\alias{mse}
\title{Accuracy/smoothness/timeliness criteria through spectral decomposition}
\usage{
mse(sweights, aweights, density = c("rw", "uniform"), passband = pi/6)
}
\arguments{
\item{sweights}{Weights of the symmetric filter (from 0 to n or -n to n).}

\item{aweights}{Weights of the asymmetric filter (from -n to m)}

\item{density}{hypothesis on the spectral density}

\item{passband}{passband threshold}
}
\value{
The criteria
}
\description{
Accuracy/smoothness/timeliness criteria through spectral decomposition
}
\examples{
filter <- filterproperties(horizon = 3, kernel = "Henderson", endpoints = "LC")
sweights <- filter$filters.coef[4:7,"q=3"]
aweights <- filter$filters.coef[1:4,"q=0"]
mse(sweights, aweights)
}
\references{
Wildi, Marc and Tucker McElroy (2019). “The trilemma between accuracy, timeliness and smoothness in real-time signal extraction”. In: International Journal of Forecasting 35.3, pp. 1072–1084.
}
