% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_filters}
\alias{plot_filters}
\alias{plot_coef}
\alias{plot_gain}
\alias{plot_phase}
\title{Plots}
\usage{
plot_coef(x, q = 0, add = FALSE, ...)

plot_gain(x, q = 0, nxlab = 6, add = FALSE, ...)

plot_phase(x, q = 0, nxlab = 6, add = FALSE, ...)
}
\arguments{
\item{x}{coefficients, gain or phase}

\item{q}{q}

\item{add}{add to previous plot}

\item{...}{other arguments to \code{plot}}

\item{nxlab}{number of xlab}
}
\value{

}
\description{
Plots
}
