% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RKHS.R
\name{rkhs_filter}
\alias{rkhs_filter}
\title{Reproducing Kernel Hilbert Space (RKHS) filters}
\usage{
rkhs_filter(
  horizon = 6,
  degree = 2,
  kernel = c("BiWeight", "Henderson", "Epanechnikov", "Triangular", "Uniform",
    "TriWeight", "TriCube"),
  asymmetricCriterion = c("Undefined", "FrequencyResponse", "Accuracy", "Smoothness",
    "Timeliness"),
  density = c("uniform", "rw"),
  passband = 2 * pi/12
)
}
\arguments{
\item{horizon}{horizon (bandwidth) of the symmetric filter .}

\item{degree}{degree of polynomial.}

\item{kernel}{kernel uses.}

\item{asymmetricCriterion}{the criteria used to compute the optimal bandwidth. If \code{"Undefined"}, \eqn{m+1} is used.}

\item{density}{hypothesis on the spectral density: \code{"uniform"} (= white woise, the default) or  \code{"rw"} (= random walk).}

\item{passband}{passband threshold.}
}
\description{
Estimation of a filter using Reproducing Kernel Hilbert Space (RKHS)
}
\examples{
rkhs <- rkhs_filter(horizon = 6, asymmetricCriterion = "Timeliness")
}
\references{
Dagum, Estela Bee and Silvia Bianconcini (2008). “The Henderson Smoother in Reproducing Kernel Hilbert Space”. In: Journal of Business & Economic Statistics 26, pp. 536–545. URL: \url{https://ideas.repec.org/a/bes/jnlbes/v26y2008p536-545.html}.
}
