% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filterproperties}
\alias{filterproperties}
\title{Get properties of local polynomials filters}
\usage{
filterproperties(
  horizon,
  degree = 3,
  kernel = c("Henderson", "Uniform", "Biweight", "Triweight", "Tricube", "Gaussian",
    "Triangular", "Parabolic"),
  endpoints = c("LC", "QL", "CQ", "CC", "DAF"),
  ic = 4.5
)
}
\arguments{
\item{horizon}{horizon of the filter}

\item{degree}{degree of polynomial}

\item{kernel}{kernel uses}

\item{endpoints}{methode for endpoints}

\item{ic}{ic ratio}
}
\value{
\code{list} with coefficients, gain and phase values
}
\description{
Get properties of local polynomials filters
}
\examples{
filterproperties(horizon = 3)
}
