% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filterproperties}
\alias{filterproperties}
\title{Title}
\usage{
filterproperties(
  horizon,
  degree = 3,
  kernel = c("Henderson", "Uniform", "Biweight", "Triweight", "Tricube", "Gaussian",
    "Triangular", "Parabolic"),
  endpoints = c("LC", "QL", "CQ", "CC", "DAF"),
  ic = 4.5,
  tweight = 0,
  passband = PI/12
)
}
\arguments{
\item{passband}{}
}
\value{

}
\description{
Title
}
