% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{fst}
\alias{fst}
\title{FST criterions}
\usage{
fst(weights, lb, passband = pi/6)
}
\arguments{
\item{weights}{Weights of the filter (from lower bound to upper bound)}

\item{lb}{Lower bound (usually negative)}

\item{passband}{Passband threshold for timeliness criterion}
}
\value{

}
\description{
FST criterions
}
\examples{
filter <- filterproperties(horizon = 6, kernel = "Henderson", endpoints = "LC")
weight <- filter$filters.coef[1:7,"q=0"]
fst(weight, lb = -6)
}
