% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{fstfilter}
\alias{fstfilter}
\title{Estimation of a filter using the Fidelity-Smoothness-Timeliness criterions}
\usage{
fstfilter(
  lags,
  leads,
  pdegree = 2,
  smoothness.weight = 1,
  smoothness.degree = 3,
  timeliness.weight = 0,
  timeliness.passband = pi/6,
  timeliness.antiphase = T
)
}
\arguments{
\item{lags}{Lags of the filter (should be positive)}

\item{leads}{Leads of the filter (should be positive or 0)}

\item{pdegree}{Local polynomials preservation: max degree}

\item{smoothness.weight}{Weight for the smoothness criterion (in \link{0, 1})}

\item{smoothness.degree}{Degree of the smoothness criterion (3 for Henderson)}

\item{timeliness.weight}{Weight for the Timeliness criterion (in [0, 1[). sweight+tweight should be in \link{0,1}}

\item{timeliness.passband}{Passband for the timeliness criterion (in radians). The phase effect is computed in \link{0, passband}}

\item{timeliness.antiphase}{See Guggemos}

\item{passband}{Passband of the filter (only used if tweight is )}
}
\value{
Contains the selected filter, its gain and phase, and the values of the 3 criterions
}
\description{
Estimation of a filter using the Fidelity-Smoothness-Timeliness criterions
}
\examples{
filter <- fstfilter(lags = 6, leads = 0)
filter$filter
filter$criterions
}
