% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_filters.R
\name{henderson}
\alias{henderson}
\title{Apply Henderson Linear Filter}
\usage{
henderson(y, length, musgrave = TRUE, ic = 4.5)
}
\arguments{
\item{y}{input time-series.}

\item{length}{length of the Henderson filter.}

\item{musgrave}{boolean indicating if musgrave asymmetric filters should be used.}

\item{ic}{ic ratio.}
}
\value{
the target signal.
}
\description{
Apply Henderson Linear Filter
}
\examples{
y <- retailsa$AllOtherGenMerchandiseStores
trend <- henderson(y, length = 13)
plot(y)
lines(trend, col = "red")
}
