% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jasym_filter.R
\name{jasym_filter}
\alias{jasym_filter}
\title{Linear Asymmetric Filtering on a Time Series}
\usage{
jasym_filter(y, coefs, lags)
}
\arguments{
\item{y}{a univariate or multivariate time series.}

\item{coefs}{an object of class \code{"moving_average"} or the coefficients
of a moving average. In that case the argument \code{lags} must be provided.}

\item{lags}{Lags of the filter (should be positive).}
}
\description{
Applies linear filtering to a univariate time series or to each series separately of a multivariate time series using an asymmetric moving average
}
\details{
For a time series \eqn{x} the asymmetric moving average is applied.
If \eqn{coef} is of length \eqn{q}, the result \eqn{y} is equal at time \eqn{t} to:
\deqn{y[t] = x[t-lags] * coef[1] + x[t-lags+1] * coef[1] + ... + x[t-lags+q] * coef[q]}
It extends the function \code{\link[stats]{filter}} that would add \code{NA} at the end of the time series.
}
\examples{
lags <- 6
leads <- 2
y <- retailsa$DrinkingPlaces
fst_coef <- fst_filter(lags = lags, leads = leads, smoothness.weight = 0.3, timeliness.weight = 0.3)
fst_coef <- fst_coef$filters.coef
lpp_coef <- lp_filter(horizon = lags, kernel = "Henderson", endpoints = "LC")
lpp_coef <- lpp_coef$filters.coef[,"q=2"]

fst_ma <- jasym_filter(y, fst_coef, lags)
lpp_ma <- jasym_filter(y, lpp_coef, lags)

plot(ts.union(y, fst_ma, lpp_ma), plot.type = "single", col = c("black","red","blue"))
}
\seealso{
\code{\link{jfilter}}
}
