% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{localpolynomials}
\alias{localpolynomials}
\title{Apply local polynomials filters}
\usage{
localpolynomials(
  y,
  horizon,
  degree = 3,
  kernel = c("Henderson", "Uniform", "Biweight", "Triweight", "Tricube", "Gaussian",
    "Triangular", "Parabolic"),
  endpoints = c("LC", "QL", "CQ", "CC", "DAF"),
  ic = 4.5
)
}
\arguments{
\item{y}{input time-series}

\item{horizon}{horizon of the filter}

\item{degree}{degree of polynomial}

\item{kernel}{kernel uses}

\item{endpoints}{methode for endpoints}

\item{ic}{ic ratio}
}
\value{

}
\description{
Apply local polynomials filters
}
