% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_moving_average.R, R/print.R
\name{moving_average}
\alias{moving_average}
\alias{is.moving_average}
\alias{is_symmetric}
\alias{upper_bound}
\alias{lower_bound}
\alias{mirror}
\alias{rev.moving_average}
\alias{length.moving_average}
\alias{to_seasonal}
\alias{sum.moving_average}
\alias{[,moving_average,numeric-method}
\alias{[,moving_average,logical-method}
\alias{[<-,moving_average,ANY,missing,numeric-method}
\alias{cbind.moving_average}
\alias{rbind.moving_average}
\alias{+,moving_average,moving_average-method}
\alias{+,moving_average,numeric-method}
\alias{+,numeric,moving_average-method}
\alias{+,moving_average,missing-method}
\alias{-,moving_average,missing-method}
\alias{-,moving_average,moving_average-method}
\alias{-,moving_average,numeric-method}
\alias{-,numeric,moving_average-method}
\alias{*,moving_average,moving_average-method}
\alias{*,moving_average,numeric-method}
\alias{*,numeric,moving_average-method}
\alias{*,ANY,moving_average-method}
\alias{*,moving_average,ANY-method}
\alias{/,moving_average,numeric-method}
\alias{^,moving_average,numeric-method}
\alias{show,moving_average-method}
\title{Manipulation of moving averages}
\usage{
moving_average(
  x,
  lags = -length(x),
  trailing_zero = FALSE,
  leading_zero = FALSE
)

is.moving_average(x)

is_symmetric(x)

upper_bound(x)

lower_bound(x)

mirror(x)

\method{rev}{moving_average}(x)

\method{length}{moving_average}(x)

to_seasonal(x, s)

\method{sum}{moving_average}(..., na.rm = FALSE)

\S4method{[}{moving_average,numeric}(x, i)

\S4method{[}{moving_average,logical}(x, i)

\S4method{[}{moving_average,ANY,missing,numeric}(x, i) <- value

\method{cbind}{moving_average}(...)

\method{rbind}{moving_average}(...)

\S4method{+}{moving_average,moving_average}(e1, e2)

\S4method{+}{moving_average,numeric}(e1, e2)

\S4method{+}{numeric,moving_average}(e1, e2)

\S4method{+}{moving_average,missing}(e1, e2)

\S4method{-}{moving_average,missing}(e1, e2)

\S4method{-}{moving_average,moving_average}(e1, e2)

\S4method{-}{moving_average,numeric}(e1, e2)

\S4method{-}{numeric,moving_average}(e1, e2)

\S4method{*}{moving_average,moving_average}(e1, e2)

\S4method{*}{moving_average,numeric}(e1, e2)

\S4method{*}{numeric,moving_average}(e1, e2)

\S4method{*}{ANY,moving_average}(e1, e2)

\S4method{*}{moving_average,ANY}(e1, e2)

\S4method{/}{moving_average,numeric}(e1, e2)

\S4method{^}{moving_average,numeric}(e1, e2)

\S4method{show}{moving_average}(object)
}
\arguments{
\item{x}{vector of coefficients}

\item{lags}{integer indicating the number of lags of the moving average.}

\item{trailing_zero, leading_zero}{boolean indicating wheter to remove leading/trailing zero and NA.}

\item{s}{seasonal period for the \code{to_seasonal()} function.}

\item{..., na.rm}{see \link[base]{mean}}

\item{i, value}{indices specifying elements to extract or replace and the new value}

\item{object, e1, e2}{moving averages}
}
\description{
Manipulation of moving averages
}
\examples{
y <- retailsa$AllOtherGenMerchandiseStores
e1 <- moving_average(rep(1,12), lags = -6)
e1 <- e1/sum(e1)
e2 <- moving_average(rep(1/12, 12), lags = -5)
M2X12 <- (e1 + e2)/2
coef(M2X12)
M3 <- moving_average(rep(1/3, 3), lags = -1)
M3X3 <- M3 * M3
# M3X3 moving average applied to each month
M3X3
M3X3_seasonal <- to_seasonal(M3X3, 12)
# M3X3_seasonal moving average applied to the global series
M3X3_seasonal

def.par <- par(no.readonly = TRUE)
par(mai = c(0.5, 0.8, 0.3, 0))
layout(matrix(c(1,2), nrow = 1))
plot_gain(M3X3, main = "M3X3 applied to each month")
plot_gain(M3X3_seasonal, main = "M3X3 applied to the global series")
par(def.par)

# To apply the moving average
t <- y * M2X12
# Or use the jfilter() function:
t <- jfilter(y, M2X12)
si <- y - t
s <- si * M3X3_seasonal
# or equivalently:
s_mm <- M3X3_seasonal * (1 - M2X12)
s <- y * s_mm
plot(s)
}
