% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mse}
\alias{mse}
\title{Accuracy/smoothness/timeliness criterions through spectral decomposition
See Wildi/McElroy}
\usage{
mse(sweights, aweights, density = c("rw", "uniform"), passband = pi/6)
}
\arguments{
\item{sweights}{Weights of the symmetric filter (from 0 to n only).}

\item{aweights}{Weights of the asymmetric filter (from -n to m)}

\item{density}{hypothesis on the spectral density}

\item{passband}{passband threshold}
}
\value{
The criterions
}
\description{
Accuracy/smoothness/timeliness criterions through spectral decomposition
See Wildi/McElroy
}
