% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_filters}
\alias{plot_filters}
\alias{plot_coef}
\alias{plot_gain}
\alias{plot_phase}
\title{Plots}
\usage{
plot_coef(
  x,
  q = 0,
  add = FALSE,
  legend = FALSE,
  legend.pos = "topright",
  zeroAsNa = FALSE,
  ...
)

plot_gain(
  x,
  q = 0,
  nxlab = 7,
  add = FALSE,
  legend = FALSE,
  legend.pos = "topright",
  xlim = c(0, pi),
  ...
)

plot_phase(
  x,
  q = 0,
  nxlab = 7,
  add = FALSE,
  legend = FALSE,
  legend.pos = "topright",
  xlim = c(0, pi),
  normalized = TRUE,
  ...
)
}
\arguments{
\item{x}{coefficients, gain or phase}

\item{q}{q}

\item{add}{boolean indicating if the new plot is added to the previous one.}

\item{legend}{boolean indicating if the legend is printed.}

\item{legend.pos}{position of the legend.}

\item{...}{other arguments to \code{matplot}}

\item{nxlab}{number of xlab}

\item{xlim}{vector containing x limits.}

\item{normalized}{boolean indicatif if the phase function is normalized by the frequency.}
}
\description{
Plots
}
\examples{
filter <- filterproperties(6, endpoints = "DAF", kernel = "Henderson")
plot_coef(filter, q = c(0,3), legend = TRUE)
plot_gain(filter, q = c(0,3), legend = TRUE)
plot_phase(filter, q = c(0,3), legend = TRUE)
}
